debug

+++ common/cfg_util.py
@@ -32,6 +32,8 @@ logger = logging.getLogger(__name__)
 
 def auth_user(user:str, t: str, cfg: dict) -> dict:
     auth_result ={"pass": False, "uid": "", "msg":""}
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         sql = f"select id, role from user where name='{user}' limit 1"
         check_info = query_sqlite(my_conn, sql)
@@ -53,6 +55,8 @@ def auth_user(user:str, t: str, cfg: dict) -> dict:
 
 def get_user_info_by_uid(uid: int)-> dict:
     user_info = {}
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             sql = f"select id, name, role, area from user where id={uid} limit 1"
@@ -73,6 +77,8 @@ def get_user_info_by_uid(uid: int)-> dict:
 def get_uid_by_user(usr_name:str) ->str:
     check_sql = f"select id from user where name='{usr_name}' limit 1"
     uid = ''
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         check_info = query_sqlite(my_conn, check_sql)
         logger.debug(f"check_info {check_info}")
@@ -89,6 +95,8 @@ def get_uid_by_user(usr_name:str) ->str:
 
 def get_user_name_by_uid(uid: int)-> str | None:
     user = None
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             sql = f"select name from user where id={uid} limit 1"
@@ -102,6 +110,8 @@ def get_user_name_by_uid(uid: int)-> str | None:
 
 def get_user_role_by_uid(uid:int)-> str | None:
     role = None
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             sql = f"select role from user where id={uid} limit 1"
@@ -115,6 +125,8 @@ def get_user_role_by_uid(uid:int)-> str | None:
 
 def get_user_hack_info(uid: int, cfg: dict)-> str | None:
     user_hack_info = None
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             sql = f"select hack_info from user where id={uid} limit 1"
@@ -142,6 +154,8 @@ def save_user_hack_info(uid: int, user_hack_info: str, cfg: dict) -> bool:
     logger.info("start_encrypt_user_hack_info")
     user_hack_info1 = encrypt(user_hack_info, cfg['sys']['cypher_key'])
     exec_sql = f"update user set hack_info ='{user_hack_info1}' where id = {uid}"
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             exec_sql = re.sub(r'\s+', ' ', exec_sql).strip()
@@ -156,6 +170,8 @@ def save_user_hack_info(uid: int, user_hack_info: str, cfg: dict) -> bool:
 
 def get_ds_cfg_by_uid(uid:int, cfg: dict) -> dict:
     config = {}
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         check_sql = (
             f"select uid, db_type, db_name, db_host, db_port, "
@@ -224,22 +240,24 @@ def save_ds_cfg(ds_cfg: dict, cfg: dict) -> bool:
                     ''')
     else:
         exec_sql = (f'''
-                    INSERT INTO db_config 
-                        (uid, db_type, db_host, db_port, db_name, db_usr, db_psw, tables, add_chart, is_strict, llm_ctx)
-                    values (
-                        '{ds_cfg["uid"]}', 
-                        '{ds_cfg["db_type"]}',
-                        '{ds_cfg["db_host"]}', 
-                        '{ds_cfg["db_port"]}', 
-                        '{ds_cfg["db_name"]}', 
-                        '{ds_cfg["db_usr_cypher"]}', 
-                        '{ds_cfg["db_psw_cypher"]}',
-                        '{ds_cfg["tables"]}',
-                        '{ds_cfg["add_chart"]}',
-                        '{ds_cfg["is_strict"]}',
-                        '{llm_ctx}'
-                    )
-                    ''')
+            INSERT INTO db_config 
+                (uid, db_type, db_host, db_port, db_name, db_usr, db_psw, tables, add_chart, is_strict, llm_ctx)
+            values (
+                '{ds_cfg["uid"]}', 
+                '{ds_cfg["db_type"]}',
+                '{ds_cfg["db_host"]}', 
+                '{ds_cfg["db_port"]}', 
+                '{ds_cfg["db_name"]}', 
+                '{ds_cfg["db_usr_cypher"]}', 
+                '{ds_cfg["db_psw_cypher"]}',
+                '{ds_cfg["tables"]}',
+                '{ds_cfg["add_chart"]}',
+                '{ds_cfg["is_strict"]}',
+                '{llm_ctx}'
+            )
+            ''')
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             exec_sql = exec_sql.replace('\n', ' ')
@@ -259,6 +277,8 @@ def save_usr(user_name: str, token: str) -> bool:
     """
     save_result = False
     exec_sql = f"INSERT INTO user (name, t) values ('{user_name}','{token}' )"
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             exec_sql = exec_sql.replace('\n', ' ')
@@ -283,6 +303,8 @@ def set_db_cache(key: str, value: str, timestamp: str, cypher_key: str) -> bool:
         raise Exception("cypher_key_null_err")
     encrypt_value = encrypt(value, cypher_key)
     exec_sql = f"INSERT INTO cache_info (key, value, timestamp) values ('{key}','{encrypt_value}', '{timestamp}')"
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             exec_sql = exec_sql.replace('\n', ' ')
@@ -303,6 +325,8 @@ def del_db_cache(key: str) -> bool:
     exec_sql = f"delete from cache_info where key='{key}'"
     if platform.system() == "Linux":
         exec_sql += " limit 1"
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             exec_sql = exec_sql.replace('\n', ' ')
@@ -322,6 +346,8 @@ def get_db_cache(key:str, cypher_key: str)->tuple | None:
     """
     if not cypher_key:
         raise Exception("cypher_key_null_err")
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             sql = f"select value ,timestamp from cache_info where key='{key}' limit 1"
@@ -347,6 +373,8 @@ def delete_data_source_config(uid: int, cfg: dict) -> bool:
     else:
         logger.error(f"no_db_source_cfg_found_for_uid_{uid}")
         return False
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         try:
             result = insert_del_sqlite(my_conn, delete_sql)
@@ -399,6 +427,8 @@ def decrypt(dt: str, key: str) -> str:
 
 def get_consts(app: str)-> dict:
     const = {}
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         sql = f"select key, value from const where app='{app}' limit 100"
         try:
@@ -412,6 +442,8 @@ def get_consts(app: str)-> dict:
 
 def get_user_list():
     user_list = []
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         sql = f"select id, name from user limit 100"
         try:
@@ -424,6 +456,8 @@ def get_user_list():
     return user_list
 
 def get_hack_info(uid: int)-> dict:
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         sql = f"select hack_q_dict from hack_list where uid = {uid} limit 1"
         try:
@@ -446,6 +480,8 @@ def get_usr_prompt_template(template_name: str,  sys_cfg: dict, uid=0)-> str:
     if not template_name or template_name == '':
         logger.warning(f"template_name_is_null_direct_return, {template_name}")
         return ""
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         sql = f"select value from prompt_template where name = '{template_name}' and  uid = {uid} limit 1"
         try:
@@ -481,6 +517,8 @@ def save_usr_prompt_template(uid: int, template_name: str, template_value: str):
     if uid == 0:
         logger.error("illegal_uid_to_set_template_err")
         return save_result
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         sql = f"select value from prompt_template where name = '{template_name}' and  uid = {uid} limit 1"
         try:
@@ -516,6 +554,8 @@ def del_usr_prompt_template(uid: int):
     if not uid or uid == 0:
         logger.error("illegal_uid_to_del_usr_template_err")
         return save_result
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         exec_sql = f"delete from prompt_template where uid = {uid} and name in ('refine_q_msg', 'sql_gen_msg')"
         try:
@@ -698,6 +738,8 @@ def save_file_info(uid: int, fid: str, full_path: str, file_suffix:int = 0) -> d
     timestamp = get_time_str()
     sql = (f"insert into file_info(uid, fid, full_path, file_suffix, timestamp) values "
            f"({uid}, '{fid}', '{full_path}', '{file_suffix}', '{timestamp}')")
+    if not os.path.exists(CFG_DB_FILE):
+        raise FileNotFoundError(f"数据库文件 {CFG_DB_FILE} 不存在")
     with sqlite3.connect(CFG_DB_FILE) as my_conn:
         logger.debug(f"save_file_info_sql, {sql}")
         my_dt = insert_del_sqlite(my_conn, sql)
